播放器进度、亮度和音量滑动控制

+++ lib/player/utils/time_format_utils.dart
@@ -8,6 +8,9 @@ class TimeFormatUtils {
 
   /// 秒数转换为分秒(分:秒)
   static String secondToMinuteAndSecond(int seconds) {
+    if (seconds < 0) {
+      return "00:00";
+    }
     int m = (seconds / 60).truncate();
     int s = seconds - m * 60;
     return "${m < 10 ? '0' : ''}$m:${s < 10 ? '0' : ''}$s";
